match types
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@21570 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ slashd/ctl_mds.h
@@ -30,8 +30,8 @@
 
 struct slmctlmsg_replpair {
 	char			scrp_addrbuf[2][RESM_ADDRBUF_SZ];
-	uint32_t		scrp_avail;
-	uint32_t		scrp_used;
+	uint64_t		scrp_avail;
+	uint64_t		scrp_used;
 };
 
 struct slmctlmsg_statfs {
